chore: 语音自动粘贴选项化 + codeloop 提示词/参数打磨 + 杂项修复
- zero-desktop/speech: 自动粘贴抽出 AutoPasteOptions（合并窗口/空格分隔/重写重打），
  原先散落的魔法值收敛为一处配置
- codeloop-core/prompt + toolkit-server/codeloop/kind: 提示词模板与解析调整
- toolkit-server: egress_sessions 补字段、shadow/stream 小修
- toolkit-worker、screenshot、proc、EgressWorkersPage: 零散修复
- Makefile.toml 任务调整
- .idea/*.iml 取消跟踪（.gitignore 已声明忽略 .idea，此文件为历史遗留）

+++ crates/zero-desktop/src/modules/screenshot/capture.rs
@@ -56,7 +56,7 @@ unsafe fn grab_rgba(rect: &MonitorRect, w: i32, h: i32) -> Result<Vec<u8>> {
         bmi.bmiHeader.biHeight = -h;
         bmi.bmiHeader.biPlanes = 1;
         bmi.bmiHeader.biBitCount = 32;
-        bmi.bmiHeader.biCompression = BI_RGB as u32;
+        bmi.bmiHeader.biCompression = BI_RGB;
 
         let stride = (w as usize) * 4;
         let mut buf = vec![0u8; stride * (h as usize)];
